fw/apps/prf/mfg_menu: split tests/extras
