+++ src/fw/apps/prf/mfg_menu.c
@@ -4,33 +4,12 @@
 #include <stdio.h>
 
 #include "applib/app.h"
-#include "applib/graphics/bitblt.h"
 #include "applib/ui/ui.h"
 #include "applib/ui/window_private.h"
-#include "applib/ui/dialogs/confirmation_dialog.h"
-#include "apps/prf/mfg_accel.h"
-#ifdef CONFIG_MAG
-#include "apps/prf/mfg_mag.h"
-#endif
-#include "apps/prf/mfg_als.h"
 #include "apps/prf/mfg_bt_device_name.h"
-#include "apps/prf/mfg_charge.h"
-#include "apps/prf/mfg_backlight.h"
-#include "apps/prf/mfg_button.h"
-#include "apps/prf/mfg_discharge.h"
-#include "apps/prf/mfg_display.h"
-#include "apps/prf/mfg_hrm.h"
-#include "apps/prf/mfg_hrm_ctr_leakage_obelix.h"
-#include "apps/prf/mfg_mic_asterix.h"
-#include "apps/prf/mfg_mic_getafix.h"
-#include "apps/prf/mfg_mic_obelix.h"
-#include "apps/prf/mfg_program_color.h"
+#include "apps/prf/mfg_extras_menu.h"
 #include "apps/prf/mfg_info_qr.h"
-#include "apps/prf/mfg_speaker_asterix.h"
-#include "apps/prf/mfg_speaker_obelix.h"
-#include "apps/prf/mfg_test_aging.h"
-#include "apps/prf/mfg_touch.h"
-#include "apps/prf/mfg_vibration.h"
+#include "apps/prf/mfg_test_menu.h"
 #include "kernel/event_loop.h"
 #include "kernel/pbl_malloc.h"
 #include "kernel/util/standby.h"
@@ -38,10 +17,8 @@
 #include "mfg/mfg_serials.h"
 #include "process_management/app_manager.h"
 #include "process_state/app_state/app_state.h"
-#include "resource/resource_ids.auto.h"
 #include "pbl/services/common/bluetooth/local_id.h"
 #include "pbl/services/common/bluetooth/pairability.h"
-#include "system/bootbits.h"
 #include "system/reset.h"
 #include "util/size.h"
 
@@ -53,20 +30,7 @@ typedef struct {
   SimpleMenuSection menu_section;
 } MfgMenuAppData;
 
-typedef struct {
-  Window *window;
-  SimpleMenuLayer *menu_layer;
-  SimpleMenuSection menu_section;
-} ExtrasMenuData;
-
 static uint16_t s_menu_position = 0;
-static bool s_show_extras_on_launch = false;
-
-#if MFG_INFO_RECORDS_TEST_RESULTS
-static GBitmap *s_menu_icons[2];
-#define ICON_IDX_CHECK 0
-#define ICON_IDX_X     1
-#endif
 
 //! Callback to run from the kernel main task
 static void prv_launch_app_cb(void *data) {
@@ -81,188 +45,12 @@ static void prv_select_info_qr(int index, void *context) {
   launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_info_qr_app_get_info());
 }
 
-static void prv_select_button(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_button_app_get_info());
-}
-
-static void prv_select_display(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_display_app_get_info());
-}
-
-#if PLATFORM_OBELIX
-static void prv_select_backlight(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_backlight_app_get_info());
-}
-#endif
-
-static void prv_select_accel(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_accel_app_get_info());
-}
-
-#ifdef CONFIG_MAG
-static void prv_select_mag(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_mag_app_get_info());
-}
-#endif
-
-static void prv_select_charge(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_charge_app_get_info());
-}
-
-static void prv_select_vibration(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_vibration_app_get_info());
-}
-
-static void prv_select_als(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_als_app_get_info());
-}
-
-#if PLATFORM_ASTERIX || PLATFORM_OBELIX
-static void prv_select_speaker(int index, void *context) {
-#if PLATFORM_ASTERIX
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_speaker_asterix_app_get_info());
-#elif PLATFORM_OBELIX
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_speaker_obelix_app_get_info());
-#endif
-}
-#endif
-
-#if PLATFORM_ASTERIX || PLATFORM_OBELIX || PLATFORM_GETAFIX
-static void prv_select_mic(int index, void *context) {
-#if PLATFORM_ASTERIX
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_mic_asterix_app_get_info());
-#elif PLATFORM_OBELIX
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_mic_obelix_app_get_info());
-#elif PLATFORM_GETAFIX
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_mic_getafix_app_get_info());
-#endif
-}
-#endif // PLATFORM_ASTERIX || PLATFORM_OBELIX || PLATFORM_GETAFIX
-
-#ifdef CONFIG_HRM
-static void prv_select_hrm(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_hrm_app_get_info());
-}
-#endif
-
-#if PLATFORM_OBELIX && defined(MANUFACTURING_FW)
-static void prv_select_hrm_ctr_leakage_obelix(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_hrm_ctr_leakage_obelix_app_get_info());
-}
-#endif
-
-#ifdef CONFIG_TOUCH
-static void prv_select_touch(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_touch_app_get_info());
-}
-#endif
-
-static void prv_select_program_color(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_program_color_app_get_info());
-}
-
-static void prv_select_test_aging(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_test_aging_app_get_info());
-}
-
-static void prv_select_discharge(int index, void *context) {
-  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_discharge_app_get_info());
-}
-
-#ifdef MANUFACTURING_FW
-static void prv_load_prf_confirmed(ClickRecognizerRef recognizer, void *context) {
-  ConfirmationDialog *confirmation_dialog = (ConfirmationDialog *)context;
-  confirmation_dialog_pop(confirmation_dialog);
-
-  bool confirmed = (click_recognizer_get_button_id(recognizer) == BUTTON_ID_UP);
-  if (confirmed) {
-    boot_bit_set(BOOT_BIT_FORCE_PRF);
-    system_reset();
-  }
-}
-
-static void prv_load_prf_click_config(void *context) {
-  window_single_click_subscribe(BUTTON_ID_UP, prv_load_prf_confirmed);
-  window_single_click_subscribe(BUTTON_ID_DOWN, prv_load_prf_confirmed);
-  window_single_click_subscribe(BUTTON_ID_BACK, prv_load_prf_confirmed);
-}
-
-static void prv_extras_select_load_prf(int index, void *context) {
-  ConfirmationDialog *confirmation_dialog = confirmation_dialog_create("Load PRF");
-  Dialog *dialog = confirmation_dialog_get_dialog(confirmation_dialog);
-
-  dialog_set_text(dialog, "Load PRF?\n\nThis action cannot be undone!");
-  dialog_set_background_color(dialog, GColorOrange);
-  dialog_set_text_color(dialog, GColorWhite);
-
-  confirmation_dialog_set_click_config_provider(confirmation_dialog, prv_load_prf_click_config);
-
-  ActionBarLayer *action_bar = confirmation_dialog_get_action_bar(confirmation_dialog);
-  action_bar_layer_set_context(action_bar, confirmation_dialog);
-
-  app_confirmation_dialog_push(confirmation_dialog);
-}
-#endif
-
-static void prv_extras_window_load(Window *window) {
-  ExtrasMenuData *data = window_get_user_data(window);
-
-  Layer *window_layer = window_get_root_layer(data->window);
-  GRect bounds = window_layer->bounds;
-
-  const SimpleMenuItem extras_menu_items[] = {
-#ifdef CONFIG_HRM
-    { .title = "Test HRM",          .callback = prv_select_hrm },
-#endif
-#ifdef MANUFACTURING_FW
-    { .title = "Load PRF",          .callback = prv_extras_select_load_prf },
-#endif
-  };
-
-  SimpleMenuItem *menu_items = app_malloc(sizeof(extras_menu_items));
-  memcpy(menu_items, extras_menu_items, sizeof(extras_menu_items));
-
-  size_t num_items = ARRAY_LENGTH(extras_menu_items);
-
-  // Add index numbers to each menu entry
-  for (size_t i = 0; i < num_items; i++) {
-    const char *original_title = menu_items[i].title;
-    size_t new_title_len = snprintf(NULL, 0, "%zu. %s", i + 1, original_title) + 1;
-    char *new_title = app_malloc(new_title_len);
-    snprintf(new_title, new_title_len, "%zu. %s", i + 1, original_title);
-    menu_items[i].title = new_title;
-  }
-
-  data->menu_section = (SimpleMenuSection) {
-    .num_items = num_items,
-    .items = menu_items
-  };
-
-  data->menu_layer = simple_menu_layer_create(bounds, data->window, &data->menu_section, 1, NULL);
-  layer_add_child(window_layer, simple_menu_layer_get_layer(data->menu_layer));
-}
-
-static void prv_extras_window_unload(Window *window) {
-  ExtrasMenuData *data = window_get_user_data(window);
-
-  simple_menu_layer_destroy(data->menu_layer);
-  app_free(data);
+static void prv_select_tests(int index, void *context) {
+  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_test_menu_app_get_info());
 }
 
 static void prv_select_extras(int index, void *context) {
-  ExtrasMenuData *data = app_malloc_check(sizeof(ExtrasMenuData));
-  *data = (ExtrasMenuData){};
-
-  data->window = window_create();
-  window_init(data->window, "Extras");
-  window_set_user_data(data->window, data);
-  window_set_window_handlers(data->window, &(WindowHandlers) {
-    .load = prv_extras_window_load,
-    .unload = prv_extras_window_unload,
-  });
-  window_set_fullscreen(data->window, true);
-
-  app_window_stack_push(data->window, true);
+  launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_extras_menu_app_get_info());
 }
 
 static void prv_select_reset(int index, void *context) {
@@ -273,81 +61,14 @@ static void prv_select_shutdown(int index, void *context) {
   enter_standby(RebootReasonCode_ShutdownMenuItem);
 }
 
-static GBitmap * prv_get_icon_for_test(MfgTest test) {
-#if MFG_INFO_RECORDS_TEST_RESULTS
-  const bool passed = mfg_info_get_test_result(test);
-  if (passed) {
-    return s_menu_icons[ICON_IDX_CHECK];
-  }
-  return s_menu_icons[ICON_IDX_X];
-#else
-  return NULL;
-#endif
-}
-
-static void prv_load_icons(void) {
-#if MFG_INFO_RECORDS_TEST_RESULTS
-  // The icons in resources are black boxes with either a white checkmark or X.
-  // In order to make them look correct in the way we are using them, we want to
-  // invert the icons so that they are black icon on a white background.
-  //
-  // To do this, load each resource temporarily and then create two new bitmaps.
-  // Then bitblt the original resource into the new bitmap using GCompOpAssignInverted.
-
-  const uint32_t icon_id[] = { RESOURCE_ID_ACTION_BAR_ICON_CHECK, RESOURCE_ID_ACTION_BAR_ICON_X };
-
-  for (unsigned i = 0; i < ARRAY_LENGTH(icon_id); ++i) {
-    GBitmap tmp;
-    gbitmap_init_with_resource(&tmp, icon_id[i]);
-
-    GBitmap *icon = gbitmap_create_blank(tmp.bounds.size, tmp.info.format);
-    bitblt_bitmap_into_bitmap(icon, &tmp, GPointZero, GCompOpAssignInverted, GColorBlack);
-
-    s_menu_icons[i] = icon;
-    gbitmap_deinit(&tmp);
-  }
-#endif
-}
-
 //! @param[out] out_items
 static size_t prv_create_menu_items(SimpleMenuItem** out_menu_items) {
-  prv_load_icons();
 
   // Define a const blueprint on the stack.
   const SimpleMenuItem s_menu_items[] = {
     { .title = "BT Device Name",    .callback = prv_select_bt_device_name },
     { .title = "Device Info",       .callback = prv_select_info_qr },
-    { .icon = prv_get_icon_for_test(MfgTest_Buttons),
-      .title = "Test Buttons",      .callback = prv_select_button },
-    { .icon = prv_get_icon_for_test(MfgTest_Display),
-      .title = "Test Display",      .callback = prv_select_display },
-#ifdef CONFIG_TOUCH
-    { .title = "Test Touch",        .callback = prv_select_touch },
-#endif
-#if PLATFORM_OBELIX
-    { .title = "Test Backlight",    .callback = prv_select_backlight },
-#endif
-    { .title = "Test Accelerometer", .callback = prv_select_accel },
-#ifdef CONFIG_MAG
-    { .title = "Test Magnetometer", .callback = prv_select_mag },
-#endif
-#if PLATFORM_ASTERIX || PLATFORM_OBELIX
-    { .title = "Test Speaker",      .callback = prv_select_speaker },
-#endif
-#if PLATFORM_ASTERIX || PLATFORM_OBELIX || PLATFORM_GETAFIX
-    { .title = "Test Microphone",   .callback = prv_select_mic },
-#endif
-    { .icon = prv_get_icon_for_test(MfgTest_ALS),
-      .title = "Test ALS",          .callback = prv_select_als },
-    { .icon = prv_get_icon_for_test(MfgTest_Vibe),
-      .title = "Test Vibration",    .callback = prv_select_vibration },
-#if PLATFORM_OBELIX && defined(MANUFACTURING_FW)
-    { .title = "Test HRM CTR/L",          .callback = prv_select_hrm_ctr_leakage_obelix },
-#endif
-    { .title = "Program Color",     .callback = prv_select_program_color },
-    { .title = "Test Aging",        .callback = prv_select_test_aging },
-    { .title = "Test Charge",       .callback = prv_select_charge },
-    { .title = "Test Discharge",    .callback = prv_select_discharge },
+    { .title = "Tests",             .callback = prv_select_tests },
     { .title = "Reset",             .callback = prv_select_reset },
     { .title = "Shutdown",          .callback = prv_select_shutdown },
     { .title = "Extras",            .callback = prv_select_extras },
@@ -379,19 +100,6 @@ static size_t prv_create_menu_items(SimpleMenuItem** out_menu_items) {
 
   (*out_menu_items)[1].subtitle = device_serial;
 
-  // Add index numbers to each menu entry
-  for (size_t i = 0; i < num_items; i++) {
-    const char *original_title = (*out_menu_items)[i].title;
-    // Allocate buffer for "N. " + original title + null terminator
-    // Max index is 99 (2 digits), so we need 4 chars for "99. " + strlen + 1
-    size_t new_title_len = snprintf(NULL, 0, "%zu. %s", i + 1, original_title) + 1;
-    char *new_title = app_malloc(new_title_len);
-    snprintf(new_title, new_title_len, "%zu. %s", i + 1, original_title);
-    (*out_menu_items)[i].title = new_title;
-  }
-
-  // We've now populated out_menu_items with the correct data. Return the number of items by
-  // looking at the original list of menu items.
   return num_items;
 }
 
@@ -423,6 +131,13 @@ static void prv_window_load(Window *window) {
 }
 
 static void s_main(void) {
+  // If returning from a submenu item, relaunch the appropriate submenu
+  if (mfg_test_menu_should_relaunch()) {
+    launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_test_menu_app_get_info());
+  } else if (mfg_extras_menu_should_relaunch()) {
+    launcher_task_add_callback(prv_launch_app_cb, (void*) mfg_extras_menu_app_get_info());
+  }
+
   bt_pairability_use();
 
   MfgMenuAppData *data = app_malloc_check(sizeof(MfgMenuAppData));
@@ -438,12 +153,6 @@ static void s_main(void) {
   window_set_overrides_back_button(data->window, true);
   window_set_fullscreen(data->window, true);
   app_window_stack_push(data->window, true /*animated*/);
-
-  // If returning from an app launched from extras menu, open extras menu
-  if (s_show_extras_on_launch) {
-    s_show_extras_on_launch = false;
-    prv_select_extras(0, NULL);
-  }
 
   app_event_loop();
